🚚 Move landing page as root page
